% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecods.R
\name{ecods}
\alias{ecods}
\title{EcoDS - ecological desk study}
\usage{
ecods(siteboundary)
}
\arguments{
\item{siteboundary}{sf object for site boundary}
}
\value{
an sf object of all sites within the buffers
}
\description{
EcoDS - ecological desk study
}
\examples{
ecods(sample_site_boundary)
}
