% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchsites.R
\name{fetch_aonb}
\alias{fetch_aonb}
\title{fetch_aonb}
\usage{
fetch_aonb(nationalbuffers)
}
\arguments{
\item{nationalbuffers}{buffer zone created by getbuffers}
}
\value{
a dataframe containing data regarding the designated site type searched
}
\description{
fetch_aonb
}
\examples{
\dontrun{aonb<-fetch_aonb(nationalbuffers)}
}
