% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchsites.R
\name{fetch_awi}
\alias{fetch_awi}
\title{Fetch Ancient Woodland data intersecting site boundary}
\usage{
fetch_awi(site)
}
\arguments{
\item{site}{sf object for site of interest, in EPSG:4326}
}
\value{
an sf object of ancient woodland data intersecting the site
}
\description{
Fetch Ancient Woodland data intersecting site boundary
}
\examples{
fetch_awi(sample_site_boundary)
}
