% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchsites.R
\name{fetch_irreplaceable}
\alias{fetch_irreplaceable}
\title{Fetch irreplaceable habitats intersecting site boundary}
\usage{
fetch_irreplaceable(site)
}
\arguments{
\item{site}{sf object for site of interest, in EPSG:4326}
}
\value{
an sf object of irreplaceable habitats intersecting the site
}
\description{
Fetch irreplaceable habitats intersecting site boundary
}
\examples{
fetch_irreplaceable(sample_site_boundary)
}
