% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchsites.R
\name{fetch_lnr}
\alias{fetch_lnr}
\title{Fetch LNRs within 2km of site}
\usage{
fetch_lnr(localbuffers)
}
\arguments{
\item{localbuffers}{buffer zone created by getbuffers}
}
\value{
a dataframe containing data regarding the designated site type searched
}
\description{
Fetch LNRs within 2km of site
}
\examples{
\dontrun{lnr<-fetch_lnr(localbuffers)}
}
