% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchsites.R
\name{fetch_lpa}
\alias{fetch_lpa}
\title{Fetch Local Planning Authority in which site is located}
\usage{
fetch_lpa(site)
}
\arguments{
\item{site}{sf object for site of interest, in EPSG:4326}
}
\value{
a string for the LPA in which the site is located
}
\description{
Fetch Local Planning Authority in which site is located
}
\examples{
fetch_lpa(sf::st_transform(sample_site_boundary, 4326))
}
