% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchsites.R
\name{fetch_natpark}
\alias{fetch_natpark}
\title{Fetch National Parks within 5km of site}
\usage{
fetch_natpark(nationalbuffers)
}
\arguments{
\item{nationalbuffers}{buffer zone created by getbuffers}
}
\value{
a dataframe containing data regarding the designated site type searched
}
\description{
Fetch National Parks within 5km of site
}
\examples{
\dontrun{natpark<-fetch_natpark(nationalbuffers)}
}
