% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchsites.R
\name{fetch_nca}
\alias{fetch_nca}
\title{Fetch National Character Area in which site is located}
\usage{
fetch_nca(site)
}
\arguments{
\item{site}{sf object for site of interest, in EPSG:4326}
}
\value{
a string for the NCA in which the site is located
}
\description{
Fetch National Character Area in which site is located
}
\examples{
fetch_nca(sample_site_boundary)
}
