% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchsites.R
\name{fetch_nnr}
\alias{fetch_nnr}
\title{fetch_nnr}
\usage{
fetch_nnr(nationalbuffers)
}
\arguments{
\item{nationalbuffers}{buffer zone created by getbuffers}
}
\value{
a dataframe containing data regarding the designated site type searched
}
\description{
fetch_nnr
}
\examples{
\dontrun{nnr<-fetch_nnr(nationalbuffers)}
}
