% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchsites.R
\name{fetch_phi}
\alias{fetch_phi}
\title{Fetch Priority Habitat Inventory (PHI) data that intersect site}
\usage{
fetch_phi(site)
}
\arguments{
\item{site}{sf object for site of interest, in EPSG:4326}
}
\value{
a dataframe containing data on priority habitats that intersect the site
}
\description{
Fetch Priority Habitat Inventory (PHI) data that intersect site
}
\examples{
\dontrun{phi<-fetch_phi(localbuffers)}
}
