% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchsites.R
\name{fetch_phi_buffer}
\alias{fetch_phi_buffer}
\title{Fetch Priority Habitat Inventory (PHI) data from 500m buffer of site}
\usage{
fetch_phi_buffer(site)
}
\arguments{
\item{site}{sf object for site of interest, in EPSG:4326}
}
\value{
a dataframe containing data on priority habitats that intersect the site
}
\description{
Fetch Priority Habitat Inventory (PHI) data from 500m buffer of site
}
\examples{
fetch_phi_buffer(sample_site_boundary)
}
