% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchsites.R
\name{fetch_phi_site}
\alias{fetch_phi_site}
\title{Fetch Priority Habitat Inventory (PHI) data that intersect site}
\usage{
fetch_phi_site(site)
}
\arguments{
\item{site}{sf object for site of interest, in EPSG:4326}
}
\value{
a dataframe containing data on priority habitats that intersect the site
}
\description{
Fetch Priority Habitat Inventory (PHI) data that intersect site
}
\examples{
fetch_phi(sample_site_boundary)
}
