% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchsites.R
\name{fetch_sac}
\alias{fetch_sac}
\title{fetch_sac}
\usage{
fetch_sac(internationalbuffers)
}
\arguments{
\item{internationalbuffers}{buffer zone created by getbuffers}
}
\value{
a dataframe containing data regarding the designated site type searched
}
\description{
fetch_sac
}
\examples{
\dontrun{sac<-fetch_sac(internationalbuffers)}
}
