% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchsites.R
\name{fetch_spa}
\alias{fetch_spa}
\title{Fetch SPAs within 10km of site}
\usage{
fetch_spa(internationalbuffers)
}
\arguments{
\item{internationalbuffers}{buffer zone created by getbuffers}
}
\value{
a dataframe containing data regarding the designated site type searched
}
\description{
Fetch SPAs within 10km of site
}
\examples{
\dontrun{spa<-fetch_spa(internationalbuffers)}
}
