% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchsites.R
\name{fetch_sssi}
\alias{fetch_sssi}
\title{fetch_sssi}
\usage{
fetch_sssi(nationalbuffers)
}
\arguments{
\item{nationalbuffers}{buffer zone created by getbuffers}
}
\value{
a dataframe containing data regarding the designated site type searched
}
\description{
fetch_sssi
}
\examples{
\dontrun{sssi<-fetch_sssi(nationalbuffers)}
}
