% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatdeskstudytables.R
\name{formatdeskstudytables}
\alias{formatdeskstudytables}
\title{formatdeskstudytables}
\usage{
formatdeskstudytables(designatedsitelist)
}
\arguments{
\item{designatedsitelist}{list containing designated sites of a given type: international, national, or local}
}
\value{
designatedsitesdf, the dataframe formatted for printing
}
\description{
formatdeskstudytables
}
\examples{
\dontrun{intnldf<-formatdeskstudytables(intnlsitelist)}
}
