% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatdsdataframe.R
\name{formatdsdataframe}
\alias{formatdsdataframe}
\title{formatdsdataframe}
\usage{
formatdsdataframe(
  dsdata,
  siteboundary,
  name_col,
  designation,
  grid_ref_col = NULL
)
}
\arguments{
\item{dsdata}{the dataframe produced by a fetch function}

\item{siteboundary}{the sf object produced by getsiteboundarydata}

\item{name_col}{the column in the desk study dataframe referring to the designated site name}

\item{designation}{the designation for the table to be formatted}

\item{grid_ref_col}{the grid ref col, if known}
}
\value{
formatted dataframe if data present, null dataframe if not
}
\description{
formatdsdataframe
}
\examples{
\dontrun{sssidf<-formatdsdataframe(sssi, siteboundary, name_col = "NAME", designation = "SSSI")}
}
