% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbuffers.R
\name{getbuffers}
\alias{getbuffers}
\title{getbuffers}
\usage{
getbuffers(siteboundary)
}
\arguments{
\item{siteboundary}{the SF object returned by getsiteboundary}
}
\value{
buffers, a list of sf objects representing the buffer zone for the desk study
}
\description{
getbuffers
}
\examples{
\dontrun{buffers<-getbuffers(siteboundary)}
}
