% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnearestdistance.R
\name{getnearestdistance}
\alias{getnearestdistance}
\title{getnearestdistance}
\usage{
getnearestdistance(designated_sites, siteboundary, name = "name")
}
\arguments{
\item{designated_sites}{the dataframe of sites produced by a given 'fetch_x' function}

\item{siteboundary}{the siteboundary object produced by the 'getsiteboundary' function}

\item{name}{the name of the column in the designated_sites DF that refers to the designated site name}
}
\value{
a dataframe containing designated site name, and the distance to the siteboundary
}
\description{
a function to be used to pull the min distances between each designated site and the site boundary
most likely to be built into the table formatting function
}
