% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsiteboundary.R
\name{getsiteboundary}
\alias{getsiteboundary}
\title{getsiteboundary}
\usage{
getsiteboundary(site_boundary_data)
}
\arguments{
\item{site_boundary_data}{the site boundary file imported initially by database loading}
}
\value{
siteboundary, a sf object formatted for desk study
}
\description{
getsiteboundary
}
\examples{
\dontrun{siteboundary<-getsiteboundary(site_boundary_data)}
}
