% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_site_boundary}
\alias{sample_site_boundary}
\title{Sample site boundary}
\format{
\subsection{\code{sample_site_boundary}}{

An sf object with 1 row and 8 columns:
\describe{
\item{id}{ID}
\item{fid}{Feature ID}
\item{site_name}{Site name}
\item{proj_code}{Project code}
\item{sbi}{Single business identifier}
\item{area_ha}{Area in hectares}
\item{created_on}{Date item created}
\item{geometry}{Geometry}
...
}
}
}
\source{
Digital Ecology Limited
}
\usage{
sample_site_boundary
}
\description{
An example site boundary spatial dataset
}
\keyword{datasets}
