% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchsites.R
\name{fetch_natpark}
\alias{fetch_natpark}
\title{fetch_natpark}
\usage{
fetch_natpark(nationalbuffers)
}
\arguments{
\item{nationalbuffers}{buffer zone created by getbuffers}
}
\value{
a dataframe containing data regarding the designated site type searched
}
\description{
fetch_natpark
}
\examples{
\dontrun{natpark<-fetch_natpark(nationalbuffers)}
}
