% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchsites.R
\name{fetch_ramsar}
\alias{fetch_ramsar}
\title{Fetch Ramsar sites within 10km of site}
\usage{
fetch_ramsar(internationalbuffers)
}
\arguments{
\item{internationalbuffers}{buffer zone created by getbuffers}
}
\value{
a dataframe containing data regarding the designated site type searched
}
\description{
Fetch Ramsar sites within 10km of site
}
\examples{
\dontrun{ramsar<-fetch_ramsar(internationalbuffers)}
}
