% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchsites.R
\name{fetch_sac}
\alias{fetch_sac}
\title{Fetch SACs within 10km of site}
\usage{
fetch_sac(internationalbuffers)
}
\arguments{
\item{internationalbuffers}{buffer zone created by getbuffers}
}
\value{
a dataframe containing data regarding the designated site type searched
}
\description{
Fetch SACs within 10km of site
}
\examples{
\dontrun{sac<-fetch_sac(internationalbuffers)}
}
