% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchsites.R
\name{fetch_spa}
\alias{fetch_spa}
\title{fetch_spa}
\usage{
fetch_spa(internationalbuffers)
}
\arguments{
\item{internationalbuffers}{buffer zone created by getbuffers}
}
\value{
a dataframe containing data regarding the designated site type searched
}
\description{
fetch_spa
}
\examples{
\dontrun{spa<-fetch_spa(internationalbuffers)}
}
