% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgridref.R
\name{getgridref}
\alias{getgridref}
\title{Identify site grid-reference}
\usage{
getgridref(sf_object)
}
\arguments{
\item{sf_object}{an sf object output from the getsiteboundary function}
}
\value{
A list containing:
\item{x_coord}{Numeric value representing the longitude of the centroid.}
\item{y_coord}{Numeric value representing the latitude of the centroid.}
\item{grid_reference_numeric}{Character string representing the numeric grid reference of the centroid.}
\item{grid_reference_letternumeric}{Character string representing the letter and numeric grid reference of the centroid.}
}
\description{
Identify site grid-reference
}
