% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgridref.R
\name{getgridref}
\alias{getgridref}
\title{Extract grid reference from coordinates of site centroid}
\usage{
getgridref(siteboundary)
}
\arguments{
\item{siteboundary}{sf object output from getsiteboundary function}
}
\value{
object with a grid reference
}
\description{
Extract grid reference from coordinates of site centroid
}
